added back some functionality that will be useful to have for independent testing
