+++ bliss/sleep.py
@@ -12,7 +12,7 @@ from torch.optim import Adam
 
 from . import device, plotting
 from .models import encoder, decoder
-from .models.encoder import get_star_bool, get_full_params, get_is_on_from_n_sources
+from .models.encoder import get_star_bool
 
 plt.switch_backend("Agg")
 
@@ -114,74 +114,6 @@ def _get_min_perm_loss(
     galaxy_bool_loss = -torch.gather(galaxy_bool_log_probs_all_perm, 1, _indx).squeeze()
 
     return locs_loss, galaxy_params_loss, star_params_loss, galaxy_bool_loss
-
-
-def get_full_params(slen: int, tile_params: dict):
-    # NOTE: off sources should have tile_locs == 0.
-    # NOTE: assume that each param in each tile is already pushed to the front.
-    required = {"n_sources", "locs"}
-    optional = {"galaxy_bool", "galaxy_params", "fluxes", "log_fluxes"}
-    assert type(slen) is int
-    assert type(tile_params) is dict
-    assert required.issubset(tile_params.keys())
-    # tile_params does not contain extraneous keys
-    for param_name in tile_params:
-        assert param_name in required or param_name in optional
-
-    tile_n_sources = tile_params["n_sources"]
-    tile_locs = tile_params["locs"]
-
-    # tile_locs shape = (n_samples x n_tiles_per_image x max_detections x 2)
-    assert len(tile_locs.shape) == 4
-    n_samples = tile_locs.shape[0]
-    n_tiles_per_image = tile_locs.shape[1]
-    max_detections = tile_locs.shape[2]
-    tile_slen = slen / math.sqrt(n_tiles_per_image)
-    n_ptiles = n_samples * n_tiles_per_image
-    assert tile_slen % 1 == 0, "Image cannot be subdivided into tiles!"
-    tile_slen = int(tile_slen)
-
-    # coordinates on tiles.
-    tile_coords = _get_tile_coords(slen, tile_slen)
-    assert tile_coords.shape[0] == n_tiles_per_image, "# tiles one image don't match"
-
-    # get is_on_array
-    tile_is_on_array_sampled = get_is_on_from_n_sources(tile_n_sources, max_detections)
-    n_sources = tile_is_on_array_sampled.sum(dim=(1, 2))  # per sample.
-    max_sources = n_sources.max().int().item()
-
-    # recenter and renormalize locations.
-    tile_is_on_array = tile_is_on_array_sampled.view(n_ptiles, -1)
-    _tile_locs = tile_locs.view(n_ptiles, -1, 2)
-    bias = tile_coords.repeat(n_samples, 1).unsqueeze(1).float()
-    _locs = (_tile_locs * tile_slen + bias) / slen
-    _locs *= tile_is_on_array.unsqueeze(2)
-
-    # sort locs and clip
-    locs = _locs.view(n_samples, -1, 2)
-    _indx_sort = _argfront(locs[..., 0], dim=1)
-    indx_sort = _indx_sort.unsqueeze(2)
-    locs = torch.gather(locs, 1, indx_sort.repeat(1, 1, 2))
-    locs = locs[:, 0:max_sources, ...]
-
-    params = {"n_sources": n_sources, "locs": locs}
-
-    # now do the same for the rest of the parameters (without scaling or biasing ofc)
-    # for same reason no need to multiply times is_on_array
-    for param_name in tile_params:
-        if param_name in optional:
-            # make sure works galaxy bool has same format as well.
-            tile_param = tile_params[param_name]
-            assert len(tile_param.shape) == 4
-            _param = tile_param.view(n_samples, n_tiles_per_image, max_detections, -1)
-            param_dim = _param.size(-1)
-            param = _param.view(n_samples, -1, param_dim)
-            param = torch.gather(param, 1, indx_sort.repeat(1, 1, param_dim))
-            param = param[:, 0:max_sources, ...]
-
-            params[param_name] = param
-
-    return params
 
 
 class SleepPhase(pl.LightningModule):
@@ -387,7 +319,7 @@ class SleepPhase(pl.LightningModule):
         )
 
     def configure_optimizers(self):
-        # TODO: use different learning rates
+        # TODO: use different optimizer parameters for each encoder.
         params = self.hparams.optimizer.params
         image_opt = Adam(self.image_encoder.parameters(), **params)
         galaxy_opt = Adam(self.galaxy_encoder.parameters(), **params)
@@ -469,7 +401,7 @@ class SleepPhase(pl.LightningModule):
         exclude = {"images", "slen", "background"}
         images = batch["images"]
         slen = int(batch["slen"].unique().item())
-        nbands = batch["images"].shape[1]
+        n_bands = batch["images"].shape[1]
         batch_size = images.shape[0]
         true_params = {k: v for k, v in batch.items() if k not in exclude}
 
@@ -519,9 +451,9 @@ class SleepPhase(pl.LightningModule):
                     locs_mse_vec.append(mse.item())
 
                 # do the same for fluxes
-                true_fluxes_i = true_fluxes[i].view(-1, nbands)
+                true_fluxes_i = true_fluxes[i].view(-1, n_bands)
                 true_fluxes_i = true_fluxes_i[: int(true_n_sources_i)]
-                fluxes_i = est_fluxes[i].view(-1, nbands)[: int(n_sources_i)]
+                fluxes_i = est_fluxes[i].view(-1, n_bands)[: int(n_sources_i)]
 
                 # sort the same way we did locations
                 true_fluxes_i = true_fluxes_i[indx_sort_true]
